Attempt to fix duplicated item names (Cross fucks up)
Uses a new strategy that appends a number to the name instead of using the GameObject's name. Also includes special handling of master rounds so that they're in order (master_round_1 - master_round_5). The console_time_stop command was removed as it was not implemented and I don't consider it needed.
Currently this may not work properly (at least on Linux/Mac) because Cross fucks up when searching for the proper static method for giving items to the player.