Fixed groups inside groups & actions on groups
Aside from the new feature of executable groups and the bug fix
that makes groups inside groups possible, this commit also includes
a new 'test' command group with a 'resources' subgroup and a 'spawn'
command. The 'spawn' command takes a single string GUID of the enemy
to spawn. The 'resources' subgroup contains only a single command,
'load', that takes at least 1 argument anduses Resources.Load to
load a resource. Arguments to it are joined with a space character.