New 'stat' group, achievements disabled by default
This commit disables achievements by default when ETGMod is installed,
use the 'conf enable_achievements true` command to enable them.
There is a new 'stat' group that can modify the TrackedStats of the save file.

The autocomplete has been improved to support different lists for different
arguments. Autocomplete now by default matches when the keyword string
is *contained* within any string in the autocompletion list, as opposed to
the any string in the list starting with it. Use the 'conf match_contains false'
command to return this behavior to how it was before this commit.