New AG&D enemies!
Found the GUIDs of all the new AG&D enemies + a few unused things.
Raccoon is still in the game and appears to still be functional. Has the
companion behavior so was supposed to be tied to an item.

Did not find the phase2 resourceful rat boss. Probably missed that as I
was sifting through the AIActor files in the asset bundle for enemies.

Playtested all these GUIDs so should be ready to be added. :D