New AG&D enemies! (#15)
* New AG&D enemies!

Found the GUIDs of all the new AG&D enemies + a few unused things.
Raccoon is still in the game and appears to still be functional. Has the
companion behavior so was supposed to be tied to an item.

Did not find the phase2 resourceful rat boss. Probably missed that as I
was sifting through the AIActor files in the asset bundle for enemies.

Playtested all these GUIDs so should be ready to be added. :D

* Revised update to enemies.txt file

Here's ALL the spawnable IDs in the game! Found by digging up enemy GUIDs from AIActor text dumps via AssetStudio on the enemies asset bundle. Found a few more things. The synergy version of Pig (the one that is green and shoots bullets) and the "revenge" version of toadie. The one that comes at you with a knife if you let him live after the Bullet King fight. He shows up on the next boss fight to come after you.

Also if you spawn the mech version of the and kill him, the game soft locks (much like the cannon boss). However spawning and killing phase one boss doesn't soft lock. Though the AI for the boss stalls after you kill him since spawning him on a floor that doesn't have the special grate he is coded to go and fall into isn't around. So he gets stuck and just does an idle animation. It doesn't prevent you from playing the game (unless maybe if you spawn him in a room that is still in combat, but you could just use tp command to exit the room I guess). 

Anyways this should cover all the new spawnable things in AG&D + a few things you missed. (like the ammoconda's little ball turret things that spawn in on occasion. Ya'll missed that one! :P ) Oh and the advanced version of the Dragun's Knife spawns I found too. :D